+++ packages/frontend/src/hooks/useEcharts.ts
@@ -26,6 +26,7 @@ import {
 import { useMemo } from 'react';
 import { useVisualizationContext } from '../components/LightdashVisualization/VisualizationProvider';
 import { getDimensionFormatter } from '../utils/resultFormatter';
+import { useOrganisation } from './organisation/useOrganisation';
 
 const getLabelFromField = (
     explore: Explore,
@@ -471,6 +472,9 @@ const useEcharts = () => {
 
     const formats = explore ? getFormats(explore) : undefined;
 
+    const { data: organisationData } = useOrganisation();
+
+
     const series = useMemo(() => {
         if (!explore || !validCartesianConfig || !resultsData) {
             return [];
@@ -541,6 +545,9 @@ const useEcharts = () => {
             top: 70, // pixels from top (makes room for legend)
             bottom: 30, // pixels from bottom (makes room for x-axis)
         },
+        ...(organisationData?.chartColors && {
+            color: organisationData?.chartColors,
+        }),
     };
 };