+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [0.47.0](https://github.com/lightdash/lightdash/compare/0.46.6...0.47.0) (2022-04-04)
+
+
+### Features
+
+* add analytics for formatted fields ([#1645](https://github.com/lightdash/lightdash/issues/1645)) ([bbcd0ba](https://github.com/lightdash/lightdash/commit/bbcd0ba6c368fda0a4518c53bb270d3dba5a6a2b))
+
+## [0.46.6](https://github.com/lightdash/lightdash/compare/0.46.5...0.46.6) (2022-04-04)
+
+
+### Bug Fixes
+
+* use echart color config ([#1639](https://github.com/lightdash/lightdash/issues/1639)) ([1869e47](https://github.com/lightdash/lightdash/commit/1869e47e194527fa5982d390cdc266df9277d2f7))
+
 ## [0.46.5](https://github.com/lightdash/lightdash/compare/0.46.4...0.46.5) (2022-04-01)